refactor: extract router logic from App component
